Added health system for Rookie and made an arraylist for rookies, to make it easier to deal with them.

+++ src/character/GameCharacter.java
@@ -5,7 +5,7 @@ public abstract class GameCharacter {
 
 
     public double x, y;
-    protected double life;
+    protected int life;
     protected double speed;
 
     protected abstract void setDestination(double x, double y);
